refactor: Split out PasswordLengthMessage into its own file
